initial connection established, app1 is a temporary solution until later verification
