+++ frontend/.env
@@ -1,1 +1,2 @@
 VITE_API_KEY="pk.eyJ1IjoiYW5zdGFuLSIsImEiOiJjbThoazl0bDQwMTgwMmxxOHRzczkxdzkyIn0.brXUg8zIfj5JksBxq8Mvqg"
+VITE_API_BASE_URL=http://localhost:5000 // URL to connect to the backend
